Added YouTube playlist support. Fixes #2

+++ commands/play.js
@@ -8,11 +8,22 @@ export const name = 'play';
 export const description = 'Adds a YouTube URL to the Music Queue.';
 export const usage = '<youtube url>';
 export async function execute(message, args) {
-  if (args[0]) {
+  const url = args[0];
+  if (url) {
     if (message.member.voiceChannel) {
       try {
-        const video = await youtube.getVideo(args[0]);
-        return handleVideo(video, message);
+        if (url.match(/^https?:\/\/(www.youtube.com|youtube.com)\/playlist(.*)$/)) {
+          const playlist = await youtube.getPlaylist(url);
+          const videos = await playlist.getVideos();
+          for (const video of Object.values(videos)) {
+            const videoId = await youtube.getVideoByID(video.id);
+            await handleVideo(videoId, message, true);
+          }
+          return message.channel.send(`✅ Playlist: **${playlist.title}** has been added to the queue!`);
+        } else {
+          const video = await youtube.getVideo(url);
+          return handleVideo(video, message);
+        }
       } catch (error) {
         console.error(error);
       }
@@ -20,7 +31,7 @@ export async function execute(message, args) {
   }
 }
 
-async function handleVideo(video, msg) {
+async function handleVideo(video, msg, playlist = false) {
   const musicQueue = msg.client.musicQueue;
   const song = {
     id: video.id,
@@ -52,6 +63,7 @@ async function handleVideo(video, msg) {
     }
   } else {
     musicQueue.songs.push(song);
+    if (playlist) return;
     msg.channel.send(`✅ **${song.title}** has been added to the queue!`);
     return await msg.delete();
   }
